tests/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0




